Refactor Generator component with VueGenerator
