+++ src/main/webapp/app/springboot/primary/vue-generator/index.ts
@@ -1,0 +1,4 @@
+import VueGeneratorComponent from './VueGenerator.component';
+import VueGeneratorVue from './VueGenerator.vue';
+
+export { VueGeneratorComponent, VueGeneratorVue };
